+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/analysis/AnalysisController.java
@@ -1,6 +1,9 @@
 package ca.corefacility.bioinformatics.irida.ria.web.analysis;
 
-import java.io.*;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.security.Principal;
 import java.util.Enumeration;
@@ -92,7 +95,7 @@ public class AnalysisController {
 	/**
 	 * Get the user {@link Analysis} list page
 	 *
-	 * @param model Model for view variables
+	 * @param model     Model for view variables
 	 * @param principal Principal {@link User}
 	 * @return Name of the analysis page view
 	 */
@@ -105,7 +108,6 @@ public class AnalysisController {
 		model.addAttribute("isAdmin", isAdmin);
 		return PAGE_ANALYSIS_LIST;
 	}
-
 
 	/**
 	 * Get the user {@link Analysis} list page
@@ -132,27 +134,9 @@ public class AnalysisController {
 
 		IridaWorkflow iridaWorkflow = workflowsService.getIridaWorkflowOrUnknown(submission);
 
-		AnalysisType analysisType = iridaWorkflow.getWorkflowDescription()
-				.getAnalysisType();
+		AnalysisType analysisType = iridaWorkflow.getWorkflowDescription().getAnalysisType();
 		model.addAttribute("analysisType", analysisType);
 		return "analysis";
-	}
-
-	/**
-	 * Get the page for viewing advanced phylogenetic visualization
-	 *
-	 * @param submissionId {@link Long} identifier for an {@link AnalysisSubmission}
-	 * @param model        {@link Model}
-	 * @return {@link String} path to the page template.
-	 */
-	@RequestMapping("/{submissionId}/advanced-phylo")
-	public String getAdvancedPhylogeneticVisualizationPage(@PathVariable Long submissionId, Model model) {
-
-		AnalysisSubmission submission = analysisSubmissionService.read(submissionId);
-
-		model.addAttribute("submissionId", submissionId);
-		model.addAttribute("submission", submission);
-		return BASE + "visualizations/phylocanvas-metadata";
 	}
 
 	/**
@@ -168,18 +152,14 @@ public class AnalysisController {
 	public void getHtmlOutputForSubmission(@PathVariable Long submissionId, @RequestParam String filename,
 			Locale locale, HttpServletResponse response) throws IOException {
 		AnalysisSubmission submission = analysisSubmissionService.read(submissionId);
-		Set<AnalysisOutputFile> files = submission.getAnalysis()
-				.getAnalysisOutputFiles();
+		Set<AnalysisOutputFile> files = submission.getAnalysis().getAnalysisOutputFiles();
 		AnalysisOutputFile outputFile = null;
 		String htmlExt = "html";
 		String htmlZipExt = "html-zip";
 		Boolean zipped = false;
 
 		for (AnalysisOutputFile file : files) {
-			if (file.getFile()
-					.toFile()
-					.getName()
-					.contains(filename)) {
+			if (file.getFile().toFile().getName().contains(filename)) {
 				String fileExt = FileUtilities.getFileExt(file.getFile());
 				if (fileExt.equals(htmlExt)) {
 					outputFile = file;
@@ -198,7 +178,7 @@ public class AnalysisController {
 		if (zipped) {
 			String htmlFile = outputFile.getFile().toFile().toString();
 			if (htmlFile.endsWith(".html.zip")) {
-				htmlFile = htmlFile.substring(0, htmlFile.length()-4);
+				htmlFile = htmlFile.substring(0, htmlFile.length() - 4);
 			}
 			try (ZipFile zipFile = new ZipFile(outputFile.getFile().toFile());
 					OutputStream outputStream = response.getOutputStream()) {
@@ -212,7 +192,7 @@ public class AnalysisController {
 				// If none match, then find the first html file in the zip
 				if (zipEntry == null) {
 					Enumeration<? extends ZipEntry> entries = zipFile.entries();
-					while(entries.hasMoreElements()) {
+					while (entries.hasMoreElements()) {
 						ZipEntry entry = entries.nextElement();
 						if (entry.getName().endsWith(".html")) {
 							zipEntry = entry;
@@ -238,8 +218,8 @@ public class AnalysisController {
 			}
 		} else {
 
-			try (InputStream inputStream = new FileInputStream(outputFile.getFile()
-					.toString()); OutputStream outputStream = response.getOutputStream()) {
+			try (InputStream inputStream = new FileInputStream(outputFile.getFile().toString());
+					OutputStream outputStream = response.getOutputStream()) {
 				// Copy the file contents to the response outputstream
 				IOUtils.copy(inputStream, outputStream);
 			} catch (IOException e) {